Added statusbar_push() to display messages in the statusbar
There are two helper functions statusbar_push that display the
text message and statusbar_pop that removes the displayed
message.

the statusbar_pop is called using g_timeout_add() and never
sould be used manually.

Note: it seems that sometimes the message could turn crazy,
so use this functions with caution :P